requirements.txt
Flask>=2.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
pytesseract>=0.3.10
easyocr>=1.7.0
pdf2image>=1.16.0
pillow-heif>=0.13.0
numpy>=1.24.0
requests>=2.31.0
geopy>=2.4.0
Werkzeug>=2.3.0
python-dotenv>=1.0.0
shapely>=2.0.0
scikit-image>=0.21.0
matplotlib>=3.7.0
flask-cors>=4.0.0 